libhns: Fix wrong type of variables and fields
Some variables an fields should be in type of unsigned instead of signed.

+++ providers/hns/hns_roce_u.h
@@ -98,7 +98,7 @@
 #define roce_set_bit(origin, shift, val) \
 	roce_set_field((origin), (1ul << (shift)), (shift), (val))
 
-#define hr_ilog32(n)		ilog32((n) - 1)
+#define hr_ilog32(n)		ilog32((unsigned int)(n) - 1)
 
 enum {
 	HNS_ROCE_QP_TABLE_BITS		= 8,
@@ -156,9 +156,9 @@ struct hns_roce_context {
 
 	pthread_mutex_t			qp_table_mutex;
 
-	int				num_qps;
-	int				qp_table_shift;
-	int				qp_table_mask;
+	uint32_t			num_qps;
+	uint32_t			qp_table_shift;
+	uint32_t			qp_table_mask;
 
 	struct hns_roce_db_page		*db_list[HNS_ROCE_DB_TYPE_NUM];
 	pthread_mutex_t			db_list_mutex;
@@ -190,7 +190,7 @@ struct hns_roce_cq {
 
 struct hns_roce_idx_que {
 	struct hns_roce_buf		buf;
-	int				entry_shift;
+	unsigned int			entry_shift;
 	unsigned long			*bitmap;
 	int				bitmap_cnt;
 };
@@ -234,7 +234,7 @@ struct hns_roce_sge_info {
 struct hns_roce_sge_ex {
 	int				offset;
 	unsigned int			sge_cnt;
-	int				sge_shift;
+	unsigned int			sge_shift;
 };
 
 struct hns_roce_rinl_sge {
